requirements.txt
# Install requirements:
# $ pip3 install -r requirements.txt

# $ pip3 install influxdb-client
influxdb-client

# $ pip3 install pyserial
pyserial
